[Refactor] domain 패키지 구조 원상복구
