+++ domain/src/main/java/edu/nextstep/camp/calculator/domain/Calculator.kt
@@ -1,4 +1,4 @@
-package edu.nextstep.camp
+package edu.nextstep.camp.calculator.domain
 
 class Calculator {
     fun calculate(rawExpression: String): Int? {
