+++ domain/src/main/java/edu/nextstep/camp/calculator/domain/Expression.kt
@@ -1,4 +1,4 @@
-package edu.nextstep.camp
+package edu.nextstep.camp.calculator.domain
 
 data class Expression(
     private val values: List<Any> = emptyList()
