+++ domain/src/main/java/edu/nextstep/camp/calculator/domain/Operator.kt
@@ -1,4 +1,4 @@
-package edu.nextstep.camp
+package edu.nextstep.camp.calculator.domain
 
 enum class Operator(
     val sign: String,
