+++ domain/src/test/kotlin/edu/nextstep/camp/calculator/domain/CalculatorTest.kt
@@ -1,6 +1,7 @@
-package edu.nextstep.camp
+package edu.nextstep.camp.calculator.domain
 
 import com.google.common.truth.Truth.assertThat
+import edu.nextstep.camp.calculator.domain.Calculator
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
