+++ domain/src/test/kotlin/edu/nextstep/camp/calculator/domain/ExpressionTest.kt
@@ -1,6 +1,8 @@
-package edu.nextstep.camp
+package edu.nextstep.camp.calculator.domain
 
 import com.google.common.truth.Truth.assertThat
+import edu.nextstep.camp.calculator.domain.Expression
+import edu.nextstep.camp.calculator.domain.Operator
 import org.junit.jupiter.api.Test
 
 class ExpressionTest {
